Rename Vitest .spec.ts files to .test.ts for naming convention consistency #147
