+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@joshuafolkken/kit",
-	"version": "0.65.0",
+	"version": "0.66.0",
 	"type": "module",
 	"license": "MIT",
 	"packageManager": "pnpm@10.33.1+sha512.05ba3c1d5d1c18f68df06470d74055e62d41fc110a0c660db1b2dfb2785327f04cf0f68345d4609bc52089e7fa0343c31593b2f9594e2c5d5da426230acc9820",
