+++ scripts-ai/completion-gate.test.ts
