+++ scripts/git/git-gh-command.test.ts
@@ -1,5 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { git_gh_command, parse_pr_state_string } from './git-gh-command'
+import { PR_CHECKS_WATCH_TIMEOUT_MS } from './git-pr-checks-watch'
 
 vi.mock('./git-gh-exec', () => ({
 	git_gh_exec: {
@@ -30,9 +31,33 @@ beforeEach(() => {
 	vi.clearAllMocks()
 })
 
+describe('git_gh_command', () => {
+	it('exposes pr_merge as a callable function', () => {
+		expect(typeof git_gh_command.pr_merge).toBe('function')
+	})
+})
+
+describe('PR_CHECKS_WATCH_TIMEOUT_MS', () => {
+	const MIN_TIMEOUT_MS = 60_000
+	const MAX_TIMEOUT_MS = 300_000
+
+	it('is defined and within a reasonable range', () => {
+		expect(PR_CHECKS_WATCH_TIMEOUT_MS).toBeGreaterThanOrEqual(MIN_TIMEOUT_MS)
+		expect(PR_CHECKS_WATCH_TIMEOUT_MS).toBeLessThanOrEqual(MAX_TIMEOUT_MS)
+	})
+})
+
 describe('parse_pr_state_string', () => {
 	it('returns trimmed string for a valid value', () => {
 		expect(parse_pr_state_string('  hello  ')).toBe('hello')
+	})
+
+	it('returns unquoted value as-is', () => {
+		expect(parse_pr_state_string('OPEN')).toBe('OPEN')
+	})
+
+	it('strips quotes and trims surrounding whitespace', () => {
+		expect(parse_pr_state_string('  "main"  ')).toBe('main')
 	})
 
 	it('returns undefined for empty string', () => {
