+++ scripts/git/git-status.test.ts
@@ -38,6 +38,21 @@ describe('git_status.list_untracked_files', () => {
 		expect(git_status.list_untracked_files('')).toStrictEqual([])
 	})
 
+	it('extracts paths from "?? " porcelain lines and ignores other states', () => {
+		const status_output = [
+			' M scripts/git/git-staging.ts',
+			'?? scripts/claude-settings.test.ts',
+			'A  src/lib/new-thing.ts',
+			'?? .claude/scheduled_tasks.lock',
+			'MM existing.ts',
+		].join('\n')
+
+		expect(git_status.list_untracked_files(status_output)).toStrictEqual([
+			'scripts/claude-settings.test.ts',
+			'.claude/scheduled_tasks.lock',
+		])
+	})
+
 	it('returns filenames for untracked lines', () => {
 		const result = git_status.list_untracked_files(UNTRACKED_LINES)
